move terrain settins config to new class
