+++ Gems/CsvSpawner/Code/Source/CsvSpawner/CsvSpawnerEditorComponent.h
@@ -12,6 +12,7 @@
 
 #include "AzFramework/Terrain/TerrainDataRequestBus.h"
 #include "CsvSpawnerUtils.h"
+#include "EditorConfigurations/CsvSpawnerEditorTerrainSettingsConfig.h"
 
 #include <AzCore/Asset/AssetCommon.h>
 #include <AzCore/Component/TickBus.h>
@@ -21,47 +22,6 @@
 
 namespace CsvSpawner
 {
-    //! Terrain Settings Configuration for Editor Component.
-    //! This config lets user decide what behaviour should be applied when the Terrain is applicable in current Level.
-    class CsvSpawnerEditorTerrainSettingsConfig
-    {
-    public:
-        AZ_RTTI(CsvSpawnerEditorTerrainSettingsConfig, CsvSpawnerEditorTerrainSettingsConfigTypeId)
-
-        CsvSpawnerEditorTerrainSettingsConfig() = default;
-        virtual ~CsvSpawnerEditorTerrainSettingsConfig() = default;
-
-        static void Reflect(AZ::ReflectContext* context);
-
-        //! Whether entities should be spawned if editor component is activated.
-        bool m_spawnOnComponentActivated{ true };
-
-        //! Prevent multiple terrains to init spawn. @returns True if spawned once, false otherwise.
-        bool m_flagSpawnEntitiesOnStartOnce{ false };
-
-        //! Whether Terrain settings or position is updated, and if should spawned entities follow up the changes.
-        bool m_spawnOnTerrainUpdate{ false };
-
-        //! Masks to be ignored while updating Terrain.
-        AzFramework::Terrain::TerrainDataNotifications::TerrainDataChangedMask m_terrainMasksToIgnore{
-            AzFramework::Terrain::TerrainDataNotifications::TerrainDataChangedMask::Settings |
-            AzFramework::Terrain::TerrainDataNotifications::TerrainDataChangedMask::ColorData
-        };
-
-    private:
-        AZ::u32 SetPropertyVisibilityByTerrain() const;
-
-        AZ::Crc32 SpawnOnTerrainUpdateTriggered();
-
-        AZ::Crc32 OnTerrainFlagsChanged();
-
-        AZ::Crc32 RefreshUI();
-
-        // Helper functions for UI Notify (need both, since cannot negate them in Reflect)
-        [[nodiscard]] bool IsSpawnOnTerrainUpdateDisabled() const;
-        [[nodiscard]] bool IsSpawnOnTerrainUpdateEnabled() const;
-    };
-
     //! Editor component for the CsvSpawner component.
     //! This component is used to spawn Csvs in the editor.
     //! It loads a CSV file that contains location of Csvs to spawn.
